fix(codex): keep MCP tests runnable without SDK in CI (#1147)
