+++ tests/test_ccr_mcp_server.py
@@ -9,7 +9,9 @@ from headroom.cache.compression_store import (
     get_compression_store,
     reset_compression_store,
 )
-from headroom.ccr import mcp_server
+from tests._mcp_stub import import_module_with_mcp_stub
+
+mcp_server = import_module_with_mcp_stub("headroom.ccr.mcp_server")
 
 
 def test_shared_stats_work_without_fcntl(monkeypatch, tmp_path) -> None:
@@ -47,7 +49,6 @@ def fresh_store():
 
 def test_mcp_uses_shared_singleton_store(fresh_store) -> None:
     """MCP's store is the global singleton, not a private instance."""
-    pytest.importorskip("mcp", reason="MCP SDK required")
     server = mcp_server.HeadroomMCPServer(check_proxy=False)
     assert server._get_local_store() is get_compression_store()
 
@@ -56,7 +57,6 @@ def test_mcp_retrieves_proxy_stored_content(fresh_store) -> None:
     """Content stored via the singleton (as the proxy does) is retrievable
     through MCP's local-store path. The HTTP fallback is disabled so this
     passes only via the shared store."""
-    pytest.importorskip("mcp", reason="MCP SDK required")
     original = '{"some": "original proxy-compressed content"}'
     hash_key = get_compression_store().store(original, '{"compressed": true}')
 
@@ -95,7 +95,6 @@ def test_mcp_retrieve_with_nonmatching_query_returns_full_content(fresh_store) -
     """A query that matches no item above the relevance floor must still return
     the stored entry (it exists and is unexpired) rather than the "Content not
     found" error, which is reserved for genuine misses."""
-    pytest.importorskip("mcp", reason="MCP SDK required")
     original = "the the the the the the the the the the\n" * 5
     hash_key = get_compression_store().store(original, "<<small>>")
     # Precondition: the query genuinely matches nothing above the BM25 floor.
@@ -112,7 +111,6 @@ def test_mcp_retrieve_with_nonmatching_query_returns_full_content(fresh_store) -
 
 def test_mcp_retrieve_missing_hash_still_errors(fresh_store) -> None:
     """A genuinely missing hash must still report "Content not found"."""
-    pytest.importorskip("mcp", reason="MCP SDK required")
     server = mcp_server.HeadroomMCPServer(check_proxy=False)
     result = asyncio.run(server._retrieve_content("nonexistent_hash", query="anything"))
     assert "Content not found" in result.get("error", "")
