+++ tests/test_memory/test_mcp_server.py
@@ -1,59 +1,13 @@
 import asyncio
-import importlib
-import sys
 from types import SimpleNamespace
 from unittest.mock import AsyncMock
 
 import numpy as np
 
 from headroom.memory.models import Memory
+from tests._mcp_stub import import_module_with_mcp_stub
 
-
-def _load_mcp_server_module():
-    if "mcp" not in sys.modules:
-        mcp_module = type(sys)("mcp")
-        mcp_server_module = type(sys)("mcp.server")
-        mcp_stdio_module = type(sys)("mcp.server.stdio")
-        mcp_types_module = type(sys)("mcp.types")
-
-        class DummyServer:
-            def __init__(self, name: str) -> None:
-                self.name = name
-
-            def list_tools(self):
-                return lambda fn: fn
-
-            def call_tool(self):
-                return lambda fn: fn
-
-            def create_initialization_options(self):
-                return {}
-
-        class DummyTool:
-            def __init__(self, **kwargs) -> None:
-                self.kwargs = kwargs
-
-        class DummyTextContent:
-            def __init__(self, **kwargs) -> None:
-                self.kwargs = kwargs
-
-        async def dummy_stdio_server():
-            raise RuntimeError("stdio_server should not run in unit tests")
-
-        mcp_server_module.Server = DummyServer
-        mcp_stdio_module.stdio_server = dummy_stdio_server
-        mcp_types_module.TextContent = DummyTextContent
-        mcp_types_module.Tool = DummyTool
-
-        sys.modules["mcp"] = mcp_module
-        sys.modules["mcp.server"] = mcp_server_module
-        sys.modules["mcp.server.stdio"] = mcp_stdio_module
-        sys.modules["mcp.types"] = mcp_types_module
-
-    return importlib.import_module("headroom.memory.mcp_server")
-
-
-mcp_server_mod = _load_mcp_server_module()
+mcp_server_mod = import_module_with_mcp_stub("headroom.memory.mcp_server")
 
 
 def test_warm_up_backend_batches_embedding_and_indexing() -> None:
